lots of QOL updates

+++ app/src/main/java/com/team695/scoutifyapp/data/repository/TaskRepository.kt
@@ -61,7 +61,7 @@ class TaskRepository(
             val gameConstants = GameConstantsStore.constants   // non-null
 
             val matchNumber = Task.matchNum
-            val teamNumber = Task.teamNum.toLong()
+            val teamNumber = Task.teamNum
             val gameType = db.matchQueries
                 .selectMatchByNumber(matchNumber)
                 .executeAsOne().gameType
